feat 实现关注操作、关注列表、粉丝列表、好友列表等功能

+++ biz/model/entity/video.go
@@ -5,12 +5,12 @@ type VideoEntity struct {
 	UserID       string `db:"user_id"`
 	Title        string `db:"title"`
 	Description  string `db:"description"`
-	CommentCount int64  `db:"comment_count"`
+	CommentCount int    `db:"comment_count"`
 	CoverURL     string `db:"cover_url"`
 	CreatedAt    string `db:"created_at"`
 	DeletedAt    string `db:"deleted_at"`
-	LikeCount    int64  `db:"like_count"`
+	LikeCount    int    `db:"like_count"`
 	UpdatedAt    string `db:"updated_at"`
 	VideoURL     string `db:"video_url"`
-	VisitCount   int64  `db:"visit_count"`
+	VisitCount   int    `db:"visit_count"`
 }
